backend/Machine_Learning/Recommendation/requirement.txt
# Function dependencies, for example:
# package>=version
google-cloud
google-cloud-aiplatform
